requirements.txt
pandas
numpy
scipy
python-levenshtein
jellyfish
networkx
segtok
elasticsearch
nltk
recordclass
unidecode
requests
Click>=6.0
click_datetime==0.2
stop-words